Move generate_injected_sorting to core.generate
