+++ src/spikeinterface/comparison/tests/test_hybrid.py
@@ -6,7 +6,6 @@ from spikeinterface.core.testing import check_recordings_equal
 from spikeinterface.comparison import (
     create_hybrid_units_recording,
     create_hybrid_spikes_recording,
-    generate_injected_sorting,
 )
 from spikeinterface.extractors import toy_example
 from spikeinterface.preprocessing import bandpass_filter
@@ -87,14 +86,6 @@ def test_hybrid_spikes_recording():
     saved_2job = hybrid_spikes_recording.save(folder=cache_folder / "spikes_2job", n_jobs=2, chunk_duration="1s")
     check_recordings_equal(hybrid_spikes_recording, saved_1job, return_scaled=False)
     check_recordings_equal(hybrid_spikes_recording, saved_2job, return_scaled=False)
-
-
-def test_generate_injected_sorting():
-    recording = load_extractor(cache_folder / "recording")
-    sorting = load_extractor(cache_folder / "sorting")
-    injected_sorting = generate_injected_sorting(
-        sorting, [recording.get_num_frames(seg_index) for seg_index in range(recording.get_num_segments())]
-    )
 
 
 if __name__ == "__main__":
