+++ src/spikeinterface/core/generate.py
@@ -177,6 +177,65 @@ def generate_sorting(
     return sorting
 
 
+def generate_injected_sorting(
+    sorting: BaseSorting,
+    num_samples: List[int],
+    max_injected_per_unit: int = 1000,
+    injected_rate: float = 0.05,
+    refractory_period_ms: float = 1.5,
+) -> NumpySorting:
+    """
+    Generates a sorting object with injected spikes from an existing sorting object.
+
+    Parameters
+    ----------
+    sorting : BaseSorting
+        The sorting object
+    num_samples : list
+        The number of samples in each segment
+    max_injected_per_unit : int, default: 1000
+        The maximum number of injected spikes per unit
+    injected_rate : float, default: 0.05
+        The rate of injected spikes with respect to the total number of spikes
+    refractory_period_ms : float, default: 1.5
+        The refractory period in ms
+
+    Returns
+    -------
+    injected_sorting : NumpySorting
+        The sorting object with injected spikes
+    """
+    injected_spike_trains = [{} for seg_index in range(sorting.get_num_segments())]
+    t_r = int(round(refractory_period_ms * sorting.get_sampling_frequency() * 1e-3))
+
+    for segment_index in range(sorting.get_num_segments()):
+        for unit_id in sorting.unit_ids:
+            spike_train = sorting.get_unit_spike_train(unit_id, segment_index=segment_index)
+            n_injection = min(max_injected_per_unit, int(round(injected_rate * len(spike_train))))
+            # Inject more, then take out all that violate the refractory period.
+            n = int(n_injection + 10 * np.sqrt(n_injection))
+            injected_spike_train = np.sort(
+                np.random.uniform(low=0, high=num_samples[segment_index], size=n).astype(np.int64)
+            )
+
+            # Remove spikes that are in the refractory period.
+            violations = np.where(np.diff(injected_spike_train) < t_r)[0]
+            injected_spike_train = np.delete(injected_spike_train, violations)
+
+            # Remove spikes that violate the refractory period of the real spikes.
+            # TODO: Need a better & faster way than this.
+            min_diff = np.min(np.abs(injected_spike_train[:, None] - spike_train[None, :]), axis=1)
+            violations = min_diff < t_r
+            injected_spike_train = injected_spike_train[~violations]
+
+            if len(injected_spike_train) > n_injection:
+                injected_spike_train = np.sort(np.random.choice(injected_spike_train, n_injection, replace=False))
+
+            injected_spike_trains[segment_index][unit_id] = injected_spike_train
+
+    return NumpySorting.from_unit_dict(injected_spike_trains, sorting.get_sampling_frequency())
+
+
 def add_synchrony_to_sorting(sorting, sync_event_ratio=0.3, seed=None):
     """
     Generates sorting object with added synchronous events from an existing sorting objects.
